requirements.txt
pymongo[srv]==4.2.0
pandas==1.5.2
scikit-learn==1.1.3
PyYAML
from-root
dill
imblearn
seaborn
matplotlib
xgboost
catboost
-e .
